fix windows compile bug

+++ Rocket/Engine/Algorism/Search.h
@@ -8,12 +8,12 @@ namespace Rocket {
         //         -- =0 T1 = T2
         //         -- <0 T1 < T2
         template<typename T>
-        int32_t BinarySearchSF(T* sorted, std::size_t size, T& key, std::function<int32_t(T&,T&)> compare) {
-            int32_t low = 0;
-            int32_t high = size - 1;
+        int64_t BinarySearchSF(T* sorted, std::size_t size, T& key, std::function<int64_t(T&,T&)> compare) {
+            int64_t low = 0;
+            int64_t high = static_cast<int64_t>(size - 1);
             while(low <= high) {
-                int32_t mid = low + (high - low) / 2;
-                int32_t result = compare(key, sorted[mid]);
+                int64_t mid = low + (high - low) / 2;
+                int64_t result = compare(key, sorted[mid]);
                 if(result < 0) high = mid - 1;
                 else if(result > 0) low = mid + 1;
                 else return mid;
@@ -23,12 +23,12 @@ namespace Rocket {
         }
         // Big first
         template<typename T>
-        int32_t BinarySearchBF(T* sorted, std::size_t size, T& key, std::function<int32_t(T&,T&)> compare) {
-            int32_t low = 0;
-            int32_t high = size - 1;
+        int64_t BinarySearchBF(T* sorted, std::size_t size, T& key, std::function<int64_t(T&,T&)> compare) {
+            int64_t low = 0;
+            int64_t high = static_cast<int64_t>(size - 1);
             while(low <= high) {
-                int32_t mid = low + (high - low) / 2;
-                int32_t result = compare(key, sorted[mid]);
+                int64_t mid = low + (high - low) / 2;
+                int64_t result = compare(key, sorted[mid]);
                 if(result > 0) high = mid - 1;
                 else if(result < 0) low = mid + 1;
                 else return mid;
