+++ UnitTest/AlgorismTest/search_test.cpp
@@ -1,13 +1,13 @@
 #include "Algorism/Search.h"
 #include <iostream>
 
-int array[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
+int64_t array[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
 
 int main() {
-    int key = 8;
+    int64_t key = 8;
 
-    auto result = Rocket::Algorism::BinarySearchSF<int>(array, sizeof(array)/sizeof(array[0]), key,
-        [](int k1, int k2) -> int32_t { return k1-k2; }
+    auto result = Rocket::Algorism::BinarySearchSF<int64_t>(array, sizeof(array)/sizeof(array[0]), key,
+        [](int64_t k1, int64_t k2) -> int64_t { return (int64_t)(k1 - k2); }
     );
 
     std::cout << "Binary Search Result: " << result << std::endl;
