+++ UnitTest/ContainerTest/bag_test.cpp
@@ -9,14 +9,14 @@ using namespace Rocket;
 int main() {
     {
         int count = 10000;
-        Bag<int32_t> bag;
+        Bag<double> bag;
         for(int i = 0; i < count; ++i) {
             bag.Add(std::rand());
         }
         auto iterator = bag.GetIterator();
         int iter_count = 0;
         while(iterator.HasNext()) {
-            [[maybe_unused]] int32_t temp = iterator.Next();
+            [[maybe_unused]] double temp = iterator.Next();
             //std::cout << "Iterator Get: " << temp << std::endl;
             iter_count++;
         }
@@ -31,7 +31,7 @@ int main() {
         auto iterator_f = bag_f.GetIterator();
         int iter_count = 0;
         while(iterator_f.HasNext()) {
-            [[maybe_unused]] int32_t temp = iterator_f.Next();
+            [[maybe_unused]] double temp = iterator_f.Next();
             iter_count++;
         }
         std::cout << "FixBag Current Size: " << bag_f.CurrentSize() << std::endl;
@@ -45,7 +45,7 @@ int main() {
         auto iterator_f = bag_l.GetIterator();
         int iter_count = 0;
         while(iterator_f.HasNext()) {
-            [[maybe_unused]] int32_t temp = iterator_f.Next();
+            [[maybe_unused]] double temp = iterator_f.Next();
             iter_count++;
         }
         std::cout << "ListBag Current Size: " << bag_l.CurrentSize() << std::endl;
